spring

+++ src/main/java/com/choong/spr/controller/MemberController.java
@@ -1,6 +1,9 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -89,23 +92,47 @@ public class MemberController {
 	}
 	
 	@GetMapping("get")
-	public void getMember(String id, Model model) {
-		MemberDto member = service.getMemberById(id);
+	public String getMember(String id, 
+			Model model, 
+			Principal principal,
+			HttpServletRequest request) {
 		
-		model.addAttribute("member", member);
+		if(hasAuthOrAdmin(id, principal, request)) {			
+			MemberDto member = service.getMemberById(id);
+			model.addAttribute("member", member);
+			
+			return null;
+		}
+		
+		return "redirect:/member/login";
+	}
+	
+	private boolean hasAuthOrAdmin(String id, Principal principal, HttpServletRequest req) {
+		return req.isUserInRole("ROLE_ADMIN") || 
+		(principal != null && principal.getName().equals(id));
 	}
 	
 	@PostMapping("remove")
-	public String removeMember(MemberDto dto, RedirectAttributes rttr) {
-		boolean success = service.removeMember(dto);
+	public String removeMember(MemberDto dto, 
+			RedirectAttributes rttr,
+			HttpServletRequest req,
+			Principal principal) {
 		
-		if(success) {
-			rttr.addFlashAttribute("message", "회원 탈퇴 되었습니다.");
-			return "redirect:/board/list";
+		if(hasAuthOrAdmin(dto.getId(), principal, req)) {
+			
+			boolean success = service.removeMember(dto);
+			
+			if(success) {
+				rttr.addFlashAttribute("message", "회원 탈퇴 되었습니다.");
+				return "redirect:/board/list";
+			}else {
+				rttr.addFlashAttribute("id", dto.getId());
+				return "redirect:/member/get";
+			}
+			
 		}else {
-			rttr.addFlashAttribute("id", dto.getId());
-			return "redirect:/member/get";
-		}		
+			return "redirect:/member/login";
+		}
 	}
 	
 	@GetMapping(path = "check2", params = "email")
@@ -133,22 +160,45 @@ public class MemberController {
 	}
 	
 	@PostMapping("modify")
-	public String modifyMember(MemberDto dto, String oldPassword, RedirectAttributes rttr) {
-		boolean success = service.modifyMember(dto, oldPassword);
+	public String modifyMember(MemberDto dto, 
+			String oldPassword, 
+			RedirectAttributes rttr,
+			HttpServletRequest req,
+			Principal principal) {
 		
-		if(success) {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
-		}else {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+		if(hasAuthOrAdmin(dto.getId(), principal, req)) {
+			
+			boolean success = service.modifyMember(dto, oldPassword);
+			
+			if(success) {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+			}else {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+			}
+			
+			rttr.addFlashAttribute("member", dto); // model object
+			rttr.addAttribute("id", dto.getId()); // query string
+			return "redirect:/member/get";
+			
+		} else {
+			return "redirect:/member/login";
 		}
-		
-		rttr.addFlashAttribute("member", dto); // model object
-		rttr.addAttribute("id", dto.getId()); // query string
-		return "redirect:/member/get"; 
 	}
 	
 	@GetMapping("login")
 	public void loginPage() {
 		
+	}
+	
+	@GetMapping("initpw")
+	public void initpwPage() {
+
+	}
+	
+	@PostMapping("initpw")
+	public String initpwPage(String id) {
+		service.initPassword(id);
+		
+		return "redirect:/board/list";
 	}
 }
